Add files via upload
This is a deep learning AI I attempted to make over the course of two days using PyTorch. As of 8-28-2023 I cannot figure out how to fix it so that it trains faster. Right now, it's not getting any better and the loss function is sitting at a massive number. I'll be trying to return to this but I'm not sure what I'm doing wrong.